Modularised code and added Kivy Carousel
Modularised code and now display images using Kivy Carousel
