+++ Appropriation Art Exhibit/.idea/misc.xml
@@ -10,5 +10,5 @@
     <ConfirmationsSetting value="0" id="Add" />
     <ConfirmationsSetting value="0" id="Remove" />
   </component>
-  <component name="ProjectRootManager" version="2" project-jdk-name="Python 2.7.10 (C:\Python27\python.exe)" project-jdk-type="Python SDK" />
+  <component name="ProjectRootManager" version="2" project-jdk-name="Python 2.7.9 (C:\Kivy-1.9.0-py2.7-win32-x64\python.bat)" project-jdk-type="Python SDK" />
 </project>
